Schema Changes for Battle
Altered the Battle_Log table to take the character id. Need to know
whose turn it is!

Added a function to get the character id who last used their turn. This
tells the application whose turn it is.

Altered the Battle_Initiate procedure to heed the Table change with the
character id.